+++ src/main/index.ts
@@ -58,3 +58,6 @@ export function drawBoxVerticalBarPlusMinus(text: string | string[][]) {
 export function drawBoxVerticalBarPlusMinusSlashes(text: string | string[][]) {
   drawBox(BoxStyles.VERTICALBAR_PLUS_MINUS_SLASHES, text);
 }
+export function drawBoxVerticalBarPlusEquals(text: string | string[][]) {
+  drawBox(BoxStyles.VERTICALBAR_PLUS_EQUALS, text);
+}